Create a fast MIME parser.
The strategy is to avoid copying data, work directly with slices of
bytes() in place. Instead of processing the message line-by-line, we use
bytes.find() to scan for subparts. We directly render our results to a
dict() which can be JSON serialized without any further processing.

For a full parse, this gives a roughly 20% performance boost for small
messages, but for large (~500k or so) messages we're over 12x faster
than the Python stdlib.

We also support lazier partial parsing, for the cases where we are only
interested in the message text or a single attachment.