Autotagging: Much progress, integration, docs, debug tools
Added `autotag-classify` command and `moggie parse --with-autotags=Y`,
for better understanding why the classifier succeeds or fails.

Attempt classification based only on special attributes first, to reduce
the impact of long, spammy messages with many words drowning out the
relatively stronger signals from the HTML classifier, subject lines, and
HeaderPrints.

In auto-training mode, do not learn from samples considered ambiguous
by the current classifier, unless we've seen the message before and it
has come up again because the user interacted with it in some way.

Added basic obfuscation to the classifier database, to avoid leaking all
the words of all the mail in the clear. Some are kept in the clear, to
allow sharing of headerprints and spammy HTML signatures.